Generic electron analyser device class added along with generic region data
